refactor MediaEntry and EditorContainer to streamline media selection handling and improve error messaging
