+++ front-end/docs/components.md
@@ -89,7 +89,7 @@ Components in `src/components` are not responsible for:
 **Responsible for**
 - screen transitions and editor flow coordination
 - media selection entry points
-- selection-time rejection messaging and flow handling based on upstream validation outcomes
+- passing selection outcomes and actions to selection UI components
 - wiring between image and GIF workflows
 - shared filter navigation and layout used across filter screens
 - shared filter controls intended to be available across media types
@@ -130,7 +130,7 @@ Top-level container for editor screen flow.
 - wiring child components together
 - passing media state and callbacks between screens
 - coordinating editor-level UI flow
-- mapping upstream media-selection validation outcomes to local selection-time UI behavior
+- mapping upstream media-selection validation outcomes to selection-flow state and callbacks
 - crop apply orchestration, including service calls and success or error UI state
 - image source decisions for crop flows, including original source versus current preview
 - persisted crop-session state that spans screens, such as the last crop box
@@ -156,6 +156,8 @@ Entry screen for starting a new edit flow.
 - upload entry points for image and video
 - camera entry point
 - local selection and validation messaging tied to this screen
+- selection-time rejection modal UI (unsupported type, size limit)
+- re-upload and retry interactions tied to image/video picker controls
 
 **Not responsible for**
 - editor flow logic beyond the initial user choice
